move the function to proper files and extract component
